pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "werkzeug>=3.1.3",
]
